[HER-170] Follow-up Rename from SeeButton into index
