+++ src/home/first-section/live/index.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Grid } from '@material-ui/core'
 
 import { useStyle } from 'home/first-section/live/hooks/use-style'
-import { SeeButton } from 'home/first-section/live/see-button/see-button'
+import { SeeButton } from 'home/first-section/live/see-button'
 
 export const Live = () => <Grid item classes={useStyle()} justify={'left'}>
   <Grid container justify={'left'}>
